Test algorithm, add debugging toggle, update pin numbers
Tested the solar tracking algorithm, confirmed that the servos move correctly in response to light changes.

Added a boolean toggle to enable debugging output as desired, rather than rewriting the code with breakpoints during tests.

Updated the pin numbers for servo control pins to pins 5 and 6, which are special pwm pins running at 980Hz instead of 490Hz. This is higher than necessary, but the higher frequency is not expected to be used for anything else so why not use these?