Create timer logic, add data printing to BT stream
Every time the main loop is ran, the system up time is checked and data is reported to bluetooth serial if the specified time interval is reached. Up time modulus the time interval is checked, and if it is fewer milliseconds than the servo move delay the data is reported. This ensures the timer stays accurate even with varying computation times.