Add speed control loop to servoInvert()
Previously the function would update the servo motors with a new position as fast as the processor could run. This resulted in the servo motors running at full speed during the function, and the function did not block so the rest of the code resumed while the servos were still moving. This causes a jittering effect where the whole assembly would get stuck and wiggle back and forth very quickly.

The function now uses the same speed control system as the main program, and features a secondary delay value in order to more precisely control the speed during the inversion action.