Test tracking code, add separate light threshold for each axis
The tracking code has been tested and is working, and the boundary handling speed control works as expected which closes #10.

The data reporting feature works, but the data is all incorrect. Closes #1, but not #11.

A separate light threshold parameter has been added so that the sensitivity of each axis can be changed independently.