Change irradiance calculation to hypothesized linear model
The calcIrrad() function was using the quadratic model from Lab 4 to determine irradiance from the photoresistors. However, this method was extremely inacurate, and the data it was made from is somehow flawed.

Implemented a linear function, with 0 volts to 0 irradiance and 2.7 volts to 400 irradiance. This model was created by guessing the optical intensity of an iPhone flashlight and should not be considered accurate, but it appears to be accurate to ~100 W/m2.