Write solar angle report function
This commit closes #7 and closes #11; however, this has not been tested and is likely innacurate because the direction of positive rotation of the servo motors and their current alignment is unknown.

Replaced placeholder value in reportEnvData() with a call to a new function calcSunPosition(), which creates a string providing the current angle of the sun according to servo motor current position.